Update Browserify & Watchify
Both browserify and watchify have changed quite a bit since the versions tracked in the current iteration of Prova.  This required a bit of an overhaul to the current way that prova interfaces with the two.  This commit updates both dependencies and modifies lib/browser and lib/browserify-transform to support the latest api.